L09. separate classes from logic.
